backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ML and Image Processing
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
pillow>=10.0.0
clip-by-openai>=1.0

# API and HTTP clients
httpx>=0.25.0
requests>=2.31.0

# Database and Caching
sqlalchemy>=2.0.0
asyncpg>=0.29.0
redis[hiredis]>=5.0.0

# Job Queue
celery>=5.3.0

# Config and Utils  
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Image utilities
numpy>=1.24.0
opencv-python>=4.8.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0